% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{print.SANvb}
\alias{print.SANvb}
\title{Print variational inference output}
\usage{
\method{print}{SANvb}(x, ...)
}
\arguments{
\item{x}{object of class \code{SANvb} (the result of a call to \code{\link{variational_CAM}}, 
\code{\link{variational_fiSAN}}, \code{\link{variational_fSAN}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The function prints a summary of the fitted model.
}
\description{
Print method for objects of class \code{SANvb}.
}
