% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SANvi-package.R
\docType{package}
\name{SANvi-package}
\alias{SANvi}
\alias{SANvi-package}
\title{SANvi: Fitting Common and Shared Atoms Nested Models via Variational Bayes}
\description{
An R package for fitting the nested common and shared atoms models using variational Bayes approximate inference for fast computation.
}
\keyword{internal}
